RF-929: Refactor login screen, move styles out
